refactor: 서버 컴포넌트에서 직접 Supabase 호출 방식으로 복원
